#23644 Update URIs
