+++ examples/iom/iom-rcs/IomToRcsExampleRunner/Program.cs
@@ -45,7 +45,7 @@ namespace IomToRcsExampleRunner
 
 				Logger.LogDebug($"Opening RCS Client from '{directoryPath}'");
 
-				var rcsClientFactory = new RcsClientFactory(directoryPath, Logger);
+				var rcsClientFactory = new RcsClientFactoryObsolete(directoryPath, Logger);
 
 				client = await rcsClientFactory.CreateRcsApiClient();
