+++ src/IdeaStatiCa.Api/RCS/RcsRestApiConstants.cs
@@ -1,0 +1,10 @@
+﻿namespace IdeaStatiCa.Api.RCS
+{
+	public class RcsRestApiConstants
+	{
+		public const string Projects = "projects";
+		public const string Sections = "sections";
+		public const string CrossSections = "cross-sections";
+		public const string InternalForces = "internal-forces";
+	}
+}
